upload updated

+++ shyira-mis/src/pages/Dashboards/logisticdashboard/fuelRequisition/viewcars.jsx
@@ -1,16 +1,19 @@
 import React, { useEffect, useState } from 'react';
 import '@fortawesome/fontawesome-free/css/all.min.css';
 import { FaEdit, FaTrash,FaTimes } from 'react-icons/fa';
+import Swal from 'sweetalert2';
 import AddNewCar from './addcarplaque'
 function CarList() {
     const [cars, setCars] = useState([]);
     const [editCarId, setEditCarId] = useState(null);
+    const [notification, setNotification] = useState('');
     const [showAddCarForm, setShowAddCarForm] = useState(false); 
     const [formData, setFormData] = useState({
       registerNumber: '',
       modeOfVehicle: '',
       dateOfReception: '',
-      depart: ''
+      depart: '',
+      destination:'',
     });
   
     useEffect(() => {
@@ -28,14 +31,38 @@ function CarList() {
     };
   
     const handleDelete = async (id) => {
+      const { value: isConfirmed } = await Swal.fire({
+  
+        title: 'Are you sure?',
+        text: "You won't be able to recover this item!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: '#3085d6',
+        confirmButtonText: 'Yes, delete it!', 
+        customClass: {
+          popup: 'custom-swal', // Apply custom class to the popup
+        }
+  
+      });
+  
+    if (isConfirmed) {
       try {
+      
         await fetch(`${process.env.REACT_APP_BACKEND_URL}/api/forms-data/cars/${id}`, {
           method: 'DELETE'
         });
+        setNotification('Car deleted successfully!'); // Set notification message
+        // Auto-remove notification after 3 seconds
+        setTimeout(() => {
+          setNotification('');
+        }, 3000);
+
         fetchCars(); // Refresh car list after deletion
       } catch (error) {
         console.error('Error deleting car:', error);
       }
+    }
     };
   
     const handleEditClick = (car) => {
@@ -44,7 +71,8 @@ function CarList() {
         registerNumber: car.registerNumber,
         modeOfVehicle: car.modeOfVehicle,
         dateOfReception: car.dateOfReception.split('T')[0], // Format date for input
-        depart: car.depart
+        depart: car.depart,
+        destination: car.destination,
       });
     };
   
@@ -75,7 +103,17 @@ function CarList() {
     return (
       <div className='veiw-car-info'>
 
+      {/* Notification Component */}
+      {notification && (
+   <div className="notification">
+       {notification}
+    </div>
 
+)}
+           <div className='add-item'>
+              <button className='add-new-user-btn' onClick={() => setShowAddCarForm(true)}>Register new car</button>
+          
+             </div>
         {editCarId ? (
           <form onSubmit={handleEditSubmit} className='addcar'>
             <h2>Edit car information</h2>
@@ -119,6 +157,16 @@ function CarList() {
                 required
               />
             </div>
+            <div>
+              <label>Destination</label>
+              <input
+                type="text"
+                name="destination"
+                value={formData.destination}
+                onChange={handleChange}
+                required
+              />
+            </div>
             <button type="submit" className='update-car'>Update Car</button>
             <button type="button" onClick={() => setEditCarId(null)} className='cancel-edit-btn'>  <FaTimes size={32} /></button>
           </form>
@@ -126,10 +174,7 @@ function CarList() {
           cars.length > 0 ? (
             <div className="car-info">
               <div className="title">
-              <div className='add-item'>
-              <button className='add-new-user-btn' onClick={() => setShowAddCarForm(true)}>Add new Item</button>
-          
-             </div>
+            
 
               <h2>Registered Cars</h2>
               </div>
@@ -143,6 +188,7 @@ function CarList() {
                   <th>Mode of Vehicle</th>
                   <th>Date of Reception</th>
                   <th>Depart</th>
+                  <th>Destination</th>
                   <th>Actions</th>
                 </tr>
               </thead>
@@ -154,6 +200,7 @@ function CarList() {
                     <td>{car.modeOfVehicle}</td>
                     <td>{new Date(car.dateOfReception).toLocaleDateString()}</td>
                     <td>{car.depart}</td>
+                    <td>{car.destination}</td>
                     <td>
                       <label htmlFor="" className='edit-icon' onClick={() => handleEditClick(car)}>
                         <i className="fas fa-edit"></i> </label>
